requirements.txt
numpy
torch
torchvision
Pillow
matplotlib
bokeh
threading
tornado
pickle
tqdm